Add new pscom benchmarks pscom_bw/pscom_bibw
This commit adds the first rudimentary versions of the pscom_bw
and pscom_bibw benchmarks for the pscom library. pscom_{,bi}bw
are modelled after the osu_bw and osu_bibw benchmarks for MPI
in the OSU benchmark suite and measure uni-directional and
bi-directional bandwidth.

The current version of the two benchmarks is limited to statically
compile-time configured parameters. At the moment bandwidth for a
single message size is reported (in units of MiBps).